refactor: build with `tshy`
