+++ .eslintrc
@@ -1,5 +1,5 @@
 {
-  "extends": ["@readme/eslint-config", "@readme/eslint-config/typescript"],
+  "extends": ["@readme/eslint-config", "@readme/eslint-config/typescript", "@readme/eslint-config/esm"],
   "root": true,
   "rules": {
     "@typescript-eslint/no-explicit-any": "off",
