+++ .tshy/esm.json
@@ -1,0 +1,11 @@
+{
+  "extends": "./build.json",
+  "include": [
+    "../src/**/*.ts",
+    "../src/**/*.mts",
+    "../src/**/*.tsx"
+  ],
+  "compilerOptions": {
+    "outDir": "../.tshy-build-tmp/esm"
+  }
+}
