+++ __tests__/__fixtures__/json-stringify-clean.ts
