+++ __tests__/lib/find-schema-definition.test.ts
@@ -1,7 +1,7 @@
 import petstore from '@readme/oas-examples/3.0/json/petstore.json';
 import { test, expect } from 'vitest';
 
-import findSchemaDefinition from '../../src/lib/find-schema-definition';
+import findSchemaDefinition from '../../src/lib/find-schema-definition.js';
 
 test('should return a definition for a given ref', () => {
   expect(findSchemaDefinition('#/components/schemas/Pet', petstore)).toStrictEqual(petstore.components.schemas.Pet);
