+++ __tests__/operation/get-parameters-as-json-schema.test.ts
@@ -1,9 +1,9 @@
-import type { OperationObject, RequestBodyObject, SchemaObject } from '../../src/rmoas.types';
+import type { OperationObject, RequestBodyObject, SchemaObject } from '../../src/rmoas.types.js';
 
 import { beforeAll, test, expect, it, describe } from 'vitest';
 
-import Oas from '../../src';
-import createOas from '../__fixtures__/create-oas';
+import Oas from '../../src/index.js';
+import createOas from '../__fixtures__/create-oas.js';
 
 let ably: Oas;
 let circular: Oas;
