+++ __tests__/operation/get-requestbody-examples.test.ts
@@ -1,7 +1,7 @@
 import { beforeAll, test, expect, it, describe } from 'vitest';
 
-import Oas from '../../src';
-import cleanStringify from '../__fixtures__/json-stringify-clean';
+import Oas from '../../src/index.js';
+import cleanStringify from '../__fixtures__/json-stringify-clean.js';
 
 let operationExamples: Oas;
 let petstore: Oas;
