+++ __tests__/samples/index.test.ts
@@ -4,11 +4,11 @@
  * @license Apache-2.0
  * @see {@link https://github.com/swagger-api/swagger-ui/blob/master/test/unit/core/plugins/samples/fn.js}
  */
-import type * as RMOAS from '../../src/rmoas.types';
+import type * as RMOAS from '../../src/rmoas.types.js';
 
 import { describe, it, expect } from 'vitest';
 
-import sampleFromSchema from '../../src/samples';
+import sampleFromSchema from '../../src/samples/index.js';
 
 describe('sampleFromSchema', () => {
   it('should be memoized', async () => {
