+++ src/lib/get-user-variable.ts
@@ -1,4 +1,4 @@
-import type * as RMOAS from '../rmoas.types';
+import type * as RMOAS from '../rmoas.types.js';
 
 /**
  * Retrieve a user variable off of a given user.
