+++ src/lib/reducer.ts
@@ -1,9 +1,9 @@
-import type { ComponentsObject, HttpMethods, OASDocument, TagObject } from '../rmoas.types';
+import type { ComponentsObject, HttpMethods, OASDocument, TagObject } from '../rmoas.types.js';
 
 import jsonPointer from 'jsonpointer';
 import { getAPIDefinitionType } from 'oas-normalize/lib/utils';
 
-import { query } from '../analyzer/util';
+import { query } from '../analyzer/util.js';
 
 export interface ReducerOptions {
   /** A key-value object of path + method combinations to reduce by. */
