+++ src/operation.ts
@@ -1,19 +1,19 @@
-import type { CallbackExamples } from './operation/get-callback-examples';
-import type { getParametersAsJSONSchemaOptions } from './operation/get-parameters-as-json-schema';
-import type { RequestBodyExamples } from './operation/get-requestbody-examples';
-import type { ResponseExamples } from './operation/get-response-examples';
+import type { CallbackExamples } from './operation/get-callback-examples.js';
+import type { getParametersAsJSONSchemaOptions } from './operation/get-parameters-as-json-schema.js';
+import type { RequestBodyExamples } from './operation/get-requestbody-examples.js';
+import type { ResponseExamples } from './operation/get-response-examples.js';
 import type { OpenAPIV3, OpenAPIV3_1 } from 'openapi-types';
 
-import dedupeCommonParameters from './lib/dedupe-common-parameters';
-import findSchemaDefinition from './lib/find-schema-definition';
-import matchesMimeType from './lib/matches-mimetype';
-import getCallbackExamples from './operation/get-callback-examples';
-import getParametersAsJSONSchema from './operation/get-parameters-as-json-schema';
-import getRequestBodyExamples from './operation/get-requestbody-examples';
-import getResponseAsJSONSchema from './operation/get-response-as-json-schema';
-import getResponseExamples from './operation/get-response-examples';
-import * as RMOAS from './rmoas.types';
-import utils from './utils';
+import dedupeCommonParameters from './lib/dedupe-common-parameters.js';
+import findSchemaDefinition from './lib/find-schema-definition.js';
+import matchesMimeType from './lib/matches-mimetype.js';
+import getCallbackExamples from './operation/get-callback-examples.js';
+import getParametersAsJSONSchema from './operation/get-parameters-as-json-schema.js';
+import getRequestBodyExamples from './operation/get-requestbody-examples.js';
+import getResponseAsJSONSchema from './operation/get-response-as-json-schema.js';
+import getResponseExamples from './operation/get-response-examples.js';
+import * as RMOAS from './rmoas.types.js';
+import utils from './utils.js';
 
 type SecurityType = 'Basic' | 'Bearer' | 'Query' | 'Header' | 'Cookie' | 'OAuth2' | 'http' | 'apiKey';
