+++ src/utils.ts
@@ -1,6 +1,6 @@
-import findSchemaDefinition from './lib/find-schema-definition';
-import matchesMimeType from './lib/matches-mimetype';
-import { types as jsonSchemaTypes } from './operation/get-parameters-as-json-schema';
+import findSchemaDefinition from './lib/find-schema-definition.js';
+import matchesMimeType from './lib/matches-mimetype.js';
+import { types as jsonSchemaTypes } from './operation/get-parameters-as-json-schema.js';
 
 const supportedMethods = new Set(['get', 'put', 'post', 'delete', 'options', 'head', 'patch', 'trace']);
